+++ App.Domain.Core/BaseData/Entities/Comment.cs
@@ -1,7 +1,9 @@
-﻿using System;
+﻿using App.Domain.Core.Product.Entities;
+using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+
+namespace App.Domain.Core.BaseData.Entities
 {
     public partial class Comment
     {
@@ -21,7 +23,7 @@ namespace App.Infrastructures.Database.SqlServer.Entities
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
 
-        public virtual Product Product { get; set; } = null!;
+        public virtual App.Domain.Core.Product.Entities.Product Product { get; set; } = null!;
         public virtual Status Status { get; set; } = null!;
         public virtual User User { get; set; } = null!;
     }
