+++ App.Domain.Core/BaseData/Entities/FileType.cs
@@ -1,7 +1,8 @@
-﻿using System;
+﻿using App.Domain.Core.Product.Entities;
+using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.BaseData.Entities
 {
     public partial class FileType
     {
