+++ App.Domain.Core/BaseData/Entities/FileTypeExtention.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.BaseData.Entities
 {
-    public partial class TagCategory
+    public partial class FileTypeExtention
     {
-        public TagCategory()
+        public FileTypeExtention()
         {
-            Tags = new HashSet<Tag>();
+            FileTypes = new HashSet<FileType>();
         }
 
         public int Id { get; set; }
@@ -15,6 +15,6 @@ namespace App.Infrastructures.Database.SqlServer.Entities
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
 
-        public virtual ICollection<Tag> Tags { get; set; }
+        public virtual ICollection<FileType> FileTypes { get; set; }
     }
 }
