+++ App.Domain.Core/Product/Entities/Brand.cs
@@ -1,18 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product.Entities
 {
-    public partial class Model
+    public partial class Brand
     {
-        public Model()
+        public Brand()
         {
             Products = new HashSet<Product>();
         }
 
         public int Id { get; set; }
-        public int ParentModelId { get; set; }
-        public int BrandId { get; set; }
+        public int DisplayOrder { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
