+++ App.Domain.Core/Product/Entities/Collection.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product.Entities
 {
     public partial class Collection
     {
