+++ App.Domain.Core/Product/Entities/Model.cs
@@ -1,17 +1,18 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product.Entities
 {
-    public partial class Brand
+    public partial class Model
     {
-        public Brand()
+        public Model()
         {
             Products = new HashSet<Product>();
         }
 
         public int Id { get; set; }
-        public int DisplayOrder { get; set; }
+        public int ParentModelId { get; set; }
+        public int BrandId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
