+++ App.Domain.Core/Product/Entities/ProductColor.cs
@@ -1,17 +1,19 @@
-﻿using System;
+﻿using App.Domain.Core.BaseData.Entities;
+using System;
 using System.Collections.Generic;
 
-namespace App.Infrastructures.Database.SqlServer.Entities
+namespace App.Domain.Core.Product.Entities
 {
-    public partial class ProductView
+    public partial class ProductColor
     {
         public int Id { get; set; }
         public int ProductId { get; set; }
-        public DateTime ViewTime { get; set; }
+        public int ColorId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime CreationDate { get; set; }
         public bool IsDeleted { get; set; }
 
+        public virtual Color Color { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
     }
 }
