+++ dotnet-eshop.sln
@@ -5,15 +5,27 @@ VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A340F267-3318-4AE1-BF7B-457C88F05385}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "01-Infrastructures", "01-Infrastructures", "{BD520F99-027D-4B73-ABB5-77BCD6C30AA5}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "02-Infrastructures", "02-Infrastructures", "{BD520F99-027D-4B73-ABB5-77BCD6C30AA5}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "02-EndPoints", "02-EndPoints", "{54355E0A-CEFB-4327-869F-BA576EE85CCC}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "03-EndPoints", "03-EndPoints", "{54355E0A-CEFB-4327-869F-BA576EE85CCC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Infrastructures.Database.SqlServer", "src\App.Infrastructures.Database.SqlServer\App.Infrastructures.Database.SqlServer.csproj", "{AC0076CE-0A1D-4402-BB1D-4BA8ABC7BCF6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "App.Infrastructures.Database.SqlServer", "src\App.Infrastructures.Database.SqlServer\App.Infrastructures.Database.SqlServer.csproj", "{AC0076CE-0A1D-4402-BB1D-4BA8ABC7BCF6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.EndPoints.Mvc.AdminUI", "src\App.EndPoints.Mvc.AdminUI\App.EndPoints.Mvc.AdminUI.csproj", "{0BE906C9-DE0E-41B9-974D-9DFFB9C72064}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "App.EndPoints.Mvc.AdminUI", "src\App.EndPoints.Mvc.AdminUI\App.EndPoints.Mvc.AdminUI.csproj", "{0BE906C9-DE0E-41B9-974D-9DFFB9C72064}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.EndPoints.Mvc.ShopUI", "src\App.EndPoints.Mvc.ShopUI\App.EndPoints.Mvc.ShopUI.csproj", "{CBF8A57D-6529-414C-9C9B-864A77961283}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "App.EndPoints.Mvc.ShopUI", "src\App.EndPoints.Mvc.ShopUI\App.EndPoints.Mvc.ShopUI.csproj", "{CBF8A57D-6529-414C-9C9B-864A77961283}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "01-Domain", "01-Domain", "{F17E874E-FC72-419F-A99E-7EDE2661B86F}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{AF92EB43-D9CE-470B-BDCA-364BA600C25D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Domain.Core", "App.Domain.Core\App.Domain.Core.csproj", "{3DC6E0AC-C95B-4943-B93D-B6E13C3B2AC5}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{F6ADD4C8-7193-442E-8D63-0848A7346EA5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Domain.Services", "App.Domain.Services\App.Domain.Services.csproj", "{19339D65-6F34-4AF1-ADAA-34DE0AAB28A3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Domain.AppServices", "App.Domain.AppServices\App.Domain.AppServices.csproj", "{29ACB00A-EE4B-4912-A0BD-AAA03F475B01}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +45,18 @@ Global
 		{CBF8A57D-6529-414C-9C9B-864A77961283}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CBF8A57D-6529-414C-9C9B-864A77961283}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CBF8A57D-6529-414C-9C9B-864A77961283}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3DC6E0AC-C95B-4943-B93D-B6E13C3B2AC5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3DC6E0AC-C95B-4943-B93D-B6E13C3B2AC5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3DC6E0AC-C95B-4943-B93D-B6E13C3B2AC5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3DC6E0AC-C95B-4943-B93D-B6E13C3B2AC5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{19339D65-6F34-4AF1-ADAA-34DE0AAB28A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19339D65-6F34-4AF1-ADAA-34DE0AAB28A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19339D65-6F34-4AF1-ADAA-34DE0AAB28A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19339D65-6F34-4AF1-ADAA-34DE0AAB28A3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{29ACB00A-EE4B-4912-A0BD-AAA03F475B01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{29ACB00A-EE4B-4912-A0BD-AAA03F475B01}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{29ACB00A-EE4B-4912-A0BD-AAA03F475B01}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{29ACB00A-EE4B-4912-A0BD-AAA03F475B01}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -43,6 +67,12 @@ Global
 		{AC0076CE-0A1D-4402-BB1D-4BA8ABC7BCF6} = {BD520F99-027D-4B73-ABB5-77BCD6C30AA5}
 		{0BE906C9-DE0E-41B9-974D-9DFFB9C72064} = {54355E0A-CEFB-4327-869F-BA576EE85CCC}
 		{CBF8A57D-6529-414C-9C9B-864A77961283} = {54355E0A-CEFB-4327-869F-BA576EE85CCC}
+		{F17E874E-FC72-419F-A99E-7EDE2661B86F} = {A340F267-3318-4AE1-BF7B-457C88F05385}
+		{AF92EB43-D9CE-470B-BDCA-364BA600C25D} = {F17E874E-FC72-419F-A99E-7EDE2661B86F}
+		{3DC6E0AC-C95B-4943-B93D-B6E13C3B2AC5} = {AF92EB43-D9CE-470B-BDCA-364BA600C25D}
+		{F6ADD4C8-7193-442E-8D63-0848A7346EA5} = {F17E874E-FC72-419F-A99E-7EDE2661B86F}
+		{19339D65-6F34-4AF1-ADAA-34DE0AAB28A3} = {F6ADD4C8-7193-442E-8D63-0848A7346EA5}
+		{29ACB00A-EE4B-4912-A0BD-AAA03F475B01} = {F6ADD4C8-7193-442E-8D63-0848A7346EA5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E33D7C88-1A57-4CA3-AE1A-7AF235D0924A}
