+++ src/App.EndPoints.Mvc.AdminUI/App.EndPoints.Mvc.AdminUI.csproj
@@ -19,6 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\App.Domain.Core\App.Domain.Core.csproj" />
     <ProjectReference Include="..\App.Infrastructures.Database.SqlServer\App.Infrastructures.Database.SqlServer.csproj" />
   </ItemGroup>
